+++ app/layouts/blocks/dashboard-hero/dashboard-hero.tsx
@@ -1,60 +1,49 @@
 import Eye from '@/assets/icons/eye.svg'
-import { accountData, bankStatementData, dashboardHeroData } from '@/data/global-data'
 import { getCurrentDate } from '@/utils/date'
 import ManWithMoney from '@/assets/images/man-w-money-ilustration.svg'
-import useLocalStorage from '@/hooks/use-local-storage'
-import { getBalanceByBankStatement } from '@/utils/bank-statement-calc'
-import Graphism from '@/assets/images/graphism.svg'
-import { useState } from 'react'
-import { twMerge } from 'tailwind-merge'
-import { IBankStatement } from '@/types/types'
-
-interface IDashboardHero {
-  amountLabel: string
-  accountLabel: string
-}
+import useStateController from '@/hooks/use-state-controller'
 
 const DashboardHero = () => {
-  const { firstName } = accountData
-  const { accountLabel, amountLabel } = dashboardHeroData as IDashboardHero
-  const { transactions } = bankStatementData as IBankStatement
-  const [isAmountVisible, setIsAmountVisible] = useState<boolean>(true)
-  const { getValue: storedBalance } = useLocalStorage('statement', transactions)
-  const calculatedBalance = getBalanceByBankStatement(storedBalance())
+  const { user, bankStatement, isLoading } = useStateController()
 
-  const balanceFormatted = new Intl.NumberFormat('pt-BR', {
+  const bankBalance = bankStatement.reduce((balance, item) => {
+    if (item.type === 'deposit') {
+      return balance + item.amount;
+    } else {
+      return balance - item.amount;
+    }
+  }, 0);
+
+   const balanceFormatted = new Intl.NumberFormat('pt-BR', {
     style: 'currency',
     currency: 'BRL',
-  }).format(calculatedBalance)
-
-  // useEffect(() => {
-  //   getValue()
-  // }, [storedBalance])
+  }).format(bankBalance)
 
   return (
-    <section className="relative flex flex-col bg-primary p-8 pb-7 pr-30 rounded-lg min-h-100 md:items-start md:flex-row sm:items-center xs:items-center overflow-hidden">
-      <Graphism className='lg:hidden absolute bottom-0 left-0 w-[9rem] md:w-[11.25rem] h-auto' />
-      <Graphism className='lg:hidden absolute top-0 right-0 w-[9rem] md:w-[11.25rem] h-auto rotate-180' />
+    <section className="flex flex-col bg-primary p-8 pb-7 pr-30 rounded-lg min-h-100 md:items-start md:flex-row sm:items-center xs:items-center overflow-hidden">
       <div className='flex-1'>
-        <h2 className="font-bold text-white text-2xl mb-6">{`Olá, ${firstName}! :)`}</h2>
+        <h5 className="font-bold text-white text-2xl mb-6">Olá, {user?.displayName}! :)</h5>
         <span className="text-white text-sm">{getCurrentDate}</span>
-        <ManWithMoney className="hidden md:block lg:hidden w-[17.6875rem] h-auto mt-12" />
+        <ManWithMoney className="md:block lg:hidden xs:hidden text-[283px] mt-12" />
       </div>
-      <div className="flex flex-col mt-10 lg:mt-24 min-w-[11.375rem]">
+      <div className="flex flex-col mt-24 ">
         <div className="flex items-center gap-6">
-          <h3 className="font-semibold text-white text-xl">{amountLabel}</h3>
-          <button className='relative' onClick={() => setIsAmountVisible(!isAmountVisible)}>
-            <span className={twMerge('absolute top-1/2 left-1/2 -translate-1/2 w-[0.125rem] h-8 rotate-45 bg-orange', isAmountVisible ? 'hidden' : 'block')} />
-            <Eye className="w-6" />
-          </button>
+          <h4 className="font-semibold text-white text-xl">Saldo</h4>
+          <Eye className="text-2xl" />
         </div>
-        <span className='w-full h-[0.125rem] bg-white lg:bg-orange my-4' />
-        <span className="text-white text-base">{accountLabel}</span>
+        <hr
+          className="mt-4 lg:bg-red md:bg-accent-text"
+          style={{
+            border: 'none',
+            height: '2px',
+          }}
+        />
+        <span className="text-white text-base mt-4">Conta corrente</span>
         <span className="font-bold text-white text-3xl mt-2">
-          {isAmountVisible ? balanceFormatted : balanceFormatted.replaceAll(/./g, '•')}
+          {isLoading ? 'Carregando...' : balanceFormatted}
         </span>
       </div>
-      <ManWithMoney className="md:hidden w-[17.6875rem] h-auto mt-12" />
+      <ManWithMoney className="md:hidden text-[283px] mt-12" />
     </section>
   )
 }
